styled-components: collapse tests into one file
