--- definitions/npm/styled-components_v4.x.x/flow_v0.75.x-/test_createGlobalStyle.js
@@ -1,12 +1,0 @@
-// @flow
-import * as React from 'react'
-import { createGlobalStyle } from 'styled-components'
-
-const GlobalStyles : React.ComponentType<*> = createGlobalStyle``
-
-const App = () =>
-  <div>
-    <GlobalStyles />
-  </div>
-
-// @NOTE not sure what better way to test this...
