--- definitions/npm/styled-components_v4.x.x/flow_v0.75.x-/test_keyframes.js
@@ -1,11 +1,0 @@
-// @flow
-import * as React from 'react'
-import {
-  keyframes,
-  KeyFrames
-} from 'styled-components'
-
-const animation : KeyFrames = keyframes``
-
-
-// @NOTE not sure what better way to test this...
