--- definitions/npm/styled-components_v4.x.x/flow_v0.75.x-/test_wrapping_classy_component.js
@@ -1,24 +1,0 @@
-// @flow
-import * as React from 'react'
-import styled from 'styled-components'
-
-// Test that we can wrap a functional component
-
-type Props = { name : string }
-
-class Hello extends React.Component<Props> {
-  render() {
-    const { name } = this.props
-
-    return (
-      <div>Hello {name}</div>
-    )
-  }
-}
-
-const StyledHello = styled(Hello)``
-
-const hello1 : React.Element<typeof Hello> = <StyledHello name="World" />
-
-// $ExpectError - Invalid prop type
-const hello2 : React.Element<typeof Hello> = <StyledHello name={3} />
